Rename to sfx + add readme + add script
