+++ client/package.json
@@ -32,6 +32,7 @@
     "@pixi/filter-glow": "^4.2.0",
     "@pixi/filter-outline": "^4.2.0",
     "@pixi/sound": "^4.4.1",
+    "audiosprite": "^0.7.2",
     "i18next": "^21.9.0",
     "labrute-fla-parser": "^1.0.12",
     "labrute-static-fla-parser": "^1.0.6",
@@ -62,6 +63,7 @@
     ]
   },
   "devDependencies": {
+    "@types/audiosprite": "^0",
     "@types/node": "^18.6.3",
     "@types/react": "^18.2.42",
     "@types/react-dom": "^18.2.17",
