+++ client/src/assets/sfx.json
@@ -1,5 +1,5 @@
 {
-  "resources": ["/sfx/mybruteaudio.mp3", "/sfx/mybruteaudio.ogg", "/sfx/mybruteaudio.m4a", "/sfx/mybruteaudio.ac3"],
+  "resources": ["/sfx/sfx.mp3", "/sfx/sfx.ogg", "/sfx/sfx.m4a", "/sfx/sfx.ac3"],
   "spritemap": {
     "poison": {
       "start": 0,
