+++ tsconfig.json
@@ -1,7 +1,6 @@
 {
   "compilerOptions": {
     "incremental": true,
-    "resolveJsonModule": true,
     "target": "es2022",
     "module": "commonjs",
     "declaration": true,
